qatar_portal/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-session==0.8.0
flask-cors==4.0.1
alembic==1.13.2
python-dotenv==1.0.1
